% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_variety.R
\name{GeomVariety}
\alias{GeomVariety}
\title{ggplot2 geom object for variety contours}
\format{
A ggplot2 \code{Geom} ggproto object.
}
\usage{
GeomVariety
}
\description{
\code{GeomVariety} powers \code{\link[=geom_variety]{geom_variety()}}.
}
\keyword{internal}
