% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_variety.R
\name{StatVariety}
\alias{StatVariety}
\title{ggplot2 stat object for variety contours}
\format{
A ggplot2 \code{Stat} ggproto object.
}
\usage{
StatVariety
}
\description{
\code{StatVariety} powers \code{\link[=stat_variety]{stat_variety()}} and \code{\link[=geom_variety]{geom_variety()}}.
}
\keyword{internal}
