% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnorm-package.R
\docType{package}
\name{vnorm}
\alias{vnorm}
\alias{vnorm-package}
\title{vnorm: Variety Normal Distribution}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Samples, projects, and visualizes variety normal distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.kahle.io/vnorm/}
  \item \url{https://github.com/dkahle/vnorm}
  \item Report bugs at \url{https://github.com/dkahle/vnorm/issues}
}

}
\author{
\strong{Maintainer}: David Kahle \email{david@kahle.io}

Authors:
\itemize{
  \item Sonish Lamsal \email{sonish_lamsal1@baylor.edu}
}

}
\keyword{internal}
