% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_stan_code.R
\name{compile_stan_code}
\alias{compile_stan_code}
\title{Compile stan model for user-defined polynomials}
\usage{
compile_stan_code(poly, custom_stan_code = FALSE, w = FALSE, homo = TRUE)
}
\arguments{
\item{poly}{An mpoly object.}

\item{custom_stan_code}{If \code{TRUE}, custom model is compiled eventhough
general case of the polynomial is already included during package installation.
Defaults to \code{FALSE}.}

\item{w}{A named list of box constraints for vectors to be passed to Stan, See
\code{\link[=rvnorm]{rvnorm()}} examples. Defauls to \code{FALSE}.}

\item{homo}{If \code{TRUE}, the sampling is done from homoskedastic variety
normal distribution. Defaulst to \code{TRUE}.

This function creates temporary stan model for the kind of polynomial user is
interested in. This also creates a variable compiled_stan_info in the Global
environmentcompiled stan models that is needed to run rvnorm using user
defined stan codes.

Usage}
}
\description{
This function helps to avoid recompiling for polynomials
with same indeterminates but different coefficients.
}
