% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_stan_code.R
\name{compile_stan_code}
\alias{compile_stan_code}
\title{Compile Stan model for user-defined polynomials}
\usage{
compile_stan_code(poly, custom_stan_code = FALSE, w = FALSE, homo = TRUE)
}
\arguments{
\item{poly}{An mpoly object.}

\item{custom_stan_code}{If \code{TRUE}, a custom model is compiled even if
the general case of the polynomial is already included during package installation.
Defaults to \code{FALSE}.}

\item{w}{A named list of box constraints for vectors to be passed to Stan. See
\code{\link[=rvnorm]{rvnorm()}} examples. Defaults to \code{FALSE}.}

\item{homo}{If \code{TRUE}, the sampling is done from a homoskedastic variety
normal distribution. Defaults to \code{TRUE}.

This function creates a temporary Stan model for the kind of polynomial the user is
interested in. It also creates a variable \code{compiled_stan_info} in the global
environment for compiled Stan models that is needed to run \code{rvnorm} using user-defined Stan codes.}
}
\description{
This function helps to avoid recompiling for polynomials
with the same indeterminate but different coefficients.
}
