% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_variety.R
\docType{data}
\name{geom_variety}
\alias{geom_variety}
\alias{stat_variety}
\alias{StatVariety}
\alias{GeomVariety}
\title{One-Dimensional Varieties in Two Dimensions}
\usage{
stat_variety(
  mapping = NULL,
  data = NULL,
  geom = GeomVariety,
  position = "identity",
  ...,
  poly = NULL,
  n = 101,
  nx = n,
  ny = n,
  xlim = NULL,
  ylim = NULL,
  shift = 0,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_variety(
  mapping = NULL,
  data = NULL,
  stat = "variety",
  position = "identity",
  ...,
  poly,
  shift = 0,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetic mappings created with \code{\link[ggplot2:aes]{ggplot2::aes()}}.}

\item{data}{Layer data. If \code{NULL}, data is inherited from the plot.}

\item{geom}{The geometric object used to display data; defaults to
\link{GeomVariety}.}

\item{position}{Position adjustment for the layer.}

\item{...}{Additional parameters forwarded to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}

\item{poly}{An \code{mpoly} or \code{mpolyList} object describing the implicit
polynomial(s) to plot.}

\item{n}{Number of grid points used in both x and y directions when \code{nx} and
\code{ny} are not supplied.}

\item{nx, ny}{Number of grid points in the x and y directions.}

\item{xlim, ylim}{Length-2 numeric vectors giving plotting limits.}

\item{shift}{A numeric constant added to the evaluated surface before
contouring, i.e. the plotted level set is \code{poly + shift = 0}.
Use this when the polynomial does not cross zero on the plotting grid.
If \code{shift = 0} and all values have one sign, \code{geom_variety()} prints a
message suggesting a shift value.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}
}
\description{
Draw implicit polynomial varieties with \code{ggplot2}.
}
\section{Aesthetics}{
 \code{\link[=geom_variety]{geom_variety()}} understands the following aesthetics
(required aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item alpha
\item color
\item group
\item linetype
\item linewidth
\item subgroup
}
}

\section{Computed variables}{

\describe{
\item{probs}{The probability associated with the highest density region,
specified
by \code{probs} argument.}
}
}

\examples{

library("ggplot2")

# 1) Ellipse
p1 <- mp("x^2 + 4 y^2 - 1")
ggplot() +
  geom_variety(poly = p1, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal()

# 2) Folium of Descartes (loop with an asymptote) - variety with singularity
p3 <- mp("x^3 + y^3 - 3 x y")
ggplot() +
  geom_variety(poly = p3, xlim = c(-2, 3), ylim = c(-2, 3)) +
  coord_equal()

# 3) "Heart" curve (classic implicit heart)
p4 <- mp("(x^2 + y^2 - 1)^3 - x^2 y^3")
ggplot() +
  geom_variety(poly = p4, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal() +
  theme(legend.position = "top")

# 4) A 2-polynomial system (mpolyList): circle and xy = 0.25
ps <- mp(c("x^2 + y^2 - 1", "x y - 0.25"))
ggplot() +
  geom_variety(poly = ps, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal()


## known issues
########################################

# 1) singularities Lemniscate of Bernoulli (figure-eight shaped)
p <- mp("(x^2 + y^2)^2 - 2 x^2")
ggplot() +
  geom_variety(poly = p, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal()

# 2) non-zero crossing components
p <- mp("x^2 + y^2 - 1")
ggplot() +
  geom_variety(poly = p, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal()

ggplot() +
  geom_variety(poly = p^2, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal()

ggplot() +
  geom_variety(
    poly = p^2, xlim = c(-2, 2), ylim = c(-2, 2), shift = -0.00101684
  ) +
  coord_equal()


}
\seealso{
\code{\link[ggplot2:geom_path]{ggplot2::geom_path()}}
}
\keyword{datasets}
