% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_variety.R
\docType{data}
\name{geom_variety}
\alias{geom_variety}
\alias{stat_variety}
\alias{StatVariety}
\alias{GeomVariety}
\title{1d Varieties in 2d}
\usage{
stat_variety(
  mapping = NULL,
  data = NULL,
  geom = GeomVariety,
  position = "identity",
  ...,
  poly = NULL,
  n = 101,
  nx = n,
  ny = n,
  xlim = NULL,
  ylim = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_variety(
  mapping = NULL,
  data = NULL,
  stat = "variety",
  position = "identity",
  ...,
  poly,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}
}
\description{
Draw a variety using the biroot and ggplot2.
}
\section{Aesthetics}{
 \code{geom_variety()} understands the following aesthetics
(required aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item alpha
\item color
\item group
\item linetype
\item linewidth
\item subgroup
}
}

\section{Computed variables}{

\describe{
\item{probs}{The probability associated with the highest density region, specified
by \code{probs} argument.}
}
}

\examples{
# Basic plots
p <- mpoly::mp("y - x")
ggplot() +
  geom_variety(poly = p, xlim = c(-2, 2), ylim = c(-2, 2))

p <- mpoly::mp("y - x^2")
ggplot() +
  geom_variety(poly = p, xlim = c(-2, 2), ylim = c(-2, 2))

p <- mpoly::mp("x^2 + y^2 - 1")
ggplot() +
  geom_variety(poly = p, xlim = c(-2, 2), ylim = c(-2, 2))

# Multiple polynomial

p1 <- mp("x^2 + y^2 - 1")
p2 <- mp("y - x")
poly <- mpolyList(p1, p2)
ggplot() +
  geom_variety(poly = poly, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal()


## Setting limits
p <- mpoly::mp("y - x^2")
ggplot() +
  geom_variety(poly = p) +
  coord_equal()

ggplot() +
  geom_variety(poly = p, xlim = c(-2, 2), ylim = c(-2, 2)) +
  coord_equal()

## ggplot2 styling
##################################################

p <- mpoly::mp("x^2 + y^2 - 1")
ggplot() +
  geom_variety(poly = p)

ggplot() +
  geom_variety(poly = p) +
  coord_equal()

ggplot() +
  geom_variety(poly = p, color = "black") +
  coord_equal()

ggplot() +
  geom_variety(poly = p, linewidth = 2) +
  coord_equal()

ggplot() +
  geom_variety(poly = p, size = 2, alpha = .2) +
  coord_equal()

ggplot() +
  geom_variety(poly = p, linetype = 2) +
  coord_equal()

ggplot() +
  geom_variety(poly = p) +
  theme_bw()

ggplot() +
  geom_variety(poly = p) +
  theme_classic()

  ggplot() +
    geom_variety(poly = mpoly::mp(c("x^2 + y^2 - 1", "(x^2 + y^2)^3 - 4 x^2 y^2")))  +
    coord_equal() + theme_void() +
    scale_color_manual(values = c("red", "blue"), guide = "none")

(p <- lissajous(7, 7, 0, 0))
ggplot() +
  geom_variety(poly = p, xlim = c(-2, 2), ylim = c(-2, 2), n = 201, show.legend = FALSE)

The above example has leend removed because the legend is too long and the
plot looks distorted. We can use theme() to put legend on top when we have
long polynomial names

p <- mp("((x + 1.5)^2 + y^2 - 1) ((x - 1.5)^2 + y^2 - 1)")
ggplot() +
  geom_variety(poly = p, xlim = c(-3,3)) +
  theme(legend.position = "top") +
  coord_equal()

## possible issues
##################################################

# at a low level, ggvariety() uses grDevices::contourLines()
# to numerically detect zero crossings. this is an imperfect process,
# so you may see gaps where none exist. as a general strategy, upping
# the number of sampled points on the grid is recommended.
# the below are commented to cut check time; they run

(p <- lissajous(7, 7, 0, 0))
ggplot() +
geom_variety(poly = p) +
  theme(legend.position = "top") +
  coord_equal()

 ggplot() +
  geom_variety(poly = p, n = 201) +
  theme(legend.position = "top") +
  coord_equal()

}
\seealso{
\code{\link[=geom_path]{geom_path()}}
}
\keyword{datasets}
