% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejection_sampler.R
\name{rejection_sampler}
\alias{rejection_sampler}
\title{Rejection sampler for variety normal distribution}
\usage{
rejection_sampler(
  n,
  poly,
  sd = 0.01,
  vars = sort(mpoly::vars(poly)),
  w = 1.25,
  output = "simple",
  dist = c("norm", "unif"),
  homo = TRUE,
  correct_p_coefficients = FALSE,
  correct_dp_coefficients = FALSE,
  message = FALSE
)
}
\arguments{
\item{n}{The number of draws desired from each chain after warmup.}

\item{poly}{An mpoly object.}

\item{sd}{The "standard deviation" component of the normal kernel.}

\item{vars}{A character vector of the indeterminates in the distribution.}

\item{w}{A named list of box constraints for vectors to be passed to Stan,
see examples. A If a single number, a box window (-w,w) is applied to all
variables.}

\item{output}{Either \code{"simple"} or \code{"tibble"} output format.}

\item{dist}{either \code{"norm"} (normal) or \code{"unif"} (uniform).}

\item{homo}{If \code{TRUE}, the sampling is done from homoskedastic variety
normal distribution.}

\item{correct_p_coefficients}{If \code{TRUE} normalize the coefficients of the polynomial.}

\item{correct_dp_coefficients}{If \code{TRUE} to normalize the coefficients of the polynomial's derivative.}

\item{message}{If \code{TRUE}, message the user about remaining numbers of sampling}
}
\value{
A matrix or tibble containing the accepted samples.
}
\description{
This function performs rejection sampling to generate samples from a
variety normal distribution.
}
