% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_stan_files.R
\name{remove_stan_files}
\alias{remove_stan_files}
\title{Remove Files Created by \code{\link[=compile_stan_code]{compile_stan_code()}}}
\usage{
remove_stan_files(path = getwd())
}
\arguments{
\item{path}{Directory used in \code{\link[=compile_stan_code]{compile_stan_code()}}.
Defaults to the current working directory.}
}
\description{
Delete Stan source and executable files tracked by the internal compile
cache created by \code{\link[=compile_stan_code]{compile_stan_code()}}.
}
