% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_stan_files.R
\name{remove_stan_files}
\alias{remove_stan_files}
\title{Remove files created by\code{\link[=compile_stan_code]{compile_stan_code()}}}
\usage{
remove_stan_file()
}
\arguments{
\item{path}{Path where \code{\link[=compile_stan_code]{compile_stan_code()}} was used and files need deletion.
Defaults to curerent working directory.}
}
\description{
Deletes all the variables and stan files created by \code{\link[=compile_stan_code]{compile_stan_code()}}
}
