% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vnorm-package.R
\docType{package}
\name{vnorm}
\alias{vnorm}
\alias{vnorm-package}
\title{vnorm: Variety Normal Distribution}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Kahle \email{david@kahle.io} (\href{https://orcid.org/0000-0002-9999-1558}{ORCID})

Authors:
\itemize{
  \item Sonish Lamsal \email{sonish_lamsal1@baylor.edu} (\href{https://orcid.org/0009-0007-1098-2211}{ORCID})
}

}
\keyword{internal}
